Rpi_Glass/requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
onnxruntime==1.14.1
requests==2.31.0
gTTS==2.5.4
picamera2==0.3.31
RPi.GPIO==0.7.1
mltu==1.0.13
